Move funs to thinc.api
